+++ epstein_files/documents/imessage/text_message.py
@@ -1,12 +1,12 @@
 import re
-from dataclasses import dataclass
+from dataclasses import dataclass, fields
 from datetime import datetime
 
 from rich.text import Text
 
 from epstein_files.util.constant.names import JEFFREY_EPSTEIN, STEVE_BANNON, UNKNOWN
 from epstein_files.util.constant.strings import TIMESTAMP_DIM
-from epstein_files.util.data import extract_last_name
+from epstein_files.util.data import extract_last_name, iso_timestamp
 from epstein_files.util.highlighted_group import get_style_for_name
 from epstein_files.util.logging import logger
 from epstein_files.util.rich import TEXT_LINK, highlighter
@@ -47,8 +47,18 @@ class TextMessage:
         if not self.id_confirmed and self.author is not None and self.author != JEFFREY_EPSTEIN:
             self.author_str += ' (?)'
 
-    def timestamp(self) -> datetime:
+    def parse_timestamp(self) -> datetime:
         return datetime.strptime(self.timestamp_str, MSG_DATE_FORMAT)
+
+    def timestamp_txt(self) -> Text:
+        timestamp_str = self.timestamp_str
+
+        try:
+            timestamp_str = iso_timestamp(self.parse_timestamp())
+        except Exception as e:
+            logger.warning(f"Failed to parse timestamp for {self}")
+
+        return Text(f"[{timestamp_str}]", style=TIMESTAMP_DIM)
 
     def _message(self) -> Text:
         lines = self.text.split('\n')
@@ -75,7 +85,24 @@ class TextMessage:
         return msg_txt
 
     def __rich__(self) -> Text:
-        timestamp_txt = Text(f"[{self.timestamp_str}]", style=TIMESTAMP_DIM).append(' ')
+        timestamp_txt = self.timestamp_txt().append(' ')
         author_style = get_style_for_name(self.author_str if self.author_str.startswith('+') else self.author)
         author_txt = Text(self.author_str, style=author_style)
         return Text('').append(timestamp_txt).append(author_txt).append(': ', style='dim').append(self._message())
+
+    def __repr__(self) -> str:
+        props = []
+        add_prop = lambda k, v: props.append(f"{k}={v}")
+
+        for _field in sorted(fields(self), key=lambda f: f.name):
+            key = _field.name
+            value = getattr(self, key)
+
+            if key == 'author_str' and self.author and self.author_str.startswith(value):
+                continue
+            elif isinstance(value, str):
+                add_prop(key, f'"{value}"')
+            else:
+                add_prop(key, value)
+
+        return f"{type(self).__name__}(" + ', '.join(props) + f')'